relayer: restructure to support subcommands
