+++ relayer/src/services/index.ts
@@ -1,4 +1,2 @@
-export * from "./arbitrum";
-export * from "./boba";
-export * from "./ethereum";
-export * from "./optimism";
+export * from "./relayer";
+export * from "./types";
